I spent most of today struggling with gettting visual studio to properly recognizer my gamdata class. In the end I decided to replace the whole project with a new copy since the file structure was messed up. Anyway I've added most of a gamedata object. Initialization seems to work. Names and Orders are both loaded into vector from text files, so feel free to add whatever names ya'll feel like. Random selection from tthe vectors is also woking. I'd like to finish up with the document generation function and start up with our inherited doctypes tomorrow.
