--- Pa9/Pa9.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Pa9", "Pa9\Pa9.vcxproj", "{7188381A-30D8-4656-A9BF-80B103E608CE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Debug|x64.ActiveCfg = Debug|x64
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Debug|x64.Build.0 = Debug|x64
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Debug|x86.ActiveCfg = Debug|Win32
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Debug|x86.Build.0 = Debug|Win32
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Release|x64.ActiveCfg = Release|x64
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Release|x64.Build.0 = Release|x64
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Release|x86.ActiveCfg = Release|Win32
-		{7188381A-30D8-4656-A9BF-80B103E608CE}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
