--- Pa9/Pa9/Information.txt
@@ -1,19 +1,0 @@
-Countries,cities
-
-"shikoma","sankaku island","gao-tse","large man","small child"
-"Redstar","gulagia","domor","garbage road","stalingrab"
-"vaustlund","megaton","gutvaster","collard","neues vegas"
-"union state","freedom land","predatoria","amunation","sealand"
-"holy city","kube"
-
-start date
-january 19 2038
-
-missions
-"movment","training","recon","spec ops"
-
-example first names
-"Noah","Liam","Mason","Jacob","William","Ethan","James","Alexander","Michael", "Benjamin"
-
-example last names
-"smith","jones","brown","johnson","williams","miller","taylor","davis","white","clark"
