--- Pa9/Pa9/corruptor.h
@@ -1,1 +1,0 @@
-#pragma once
