+++ paperclone.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2005
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "paperclone", "paperclone\paperclone.vcxproj", "{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Debug|x64.ActiveCfg = Debug|x64
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Debug|x64.Build.0 = Debug|x64
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Debug|x86.ActiveCfg = Debug|Win32
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Debug|x86.Build.0 = Debug|Win32
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Release|x64.ActiveCfg = Release|x64
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Release|x64.Build.0 = Release|x64
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Release|x86.ActiveCfg = Release|Win32
+		{C8368200-D791-4AFD-BBFC-E7EDEDD436DA}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0C6B0FF4-B5F7-4134-9C10-A1D9623ABE6B}
+	EndGlobalSection
+EndGlobal
