+++ paperclone/Debug/paperclone.log
@@ -1,0 +1,2 @@
+﻿  gamedata.cpp
+  paperclone.vcxproj -> C:\Users\COOL USER\source\repos\Papes Pls\Debug\paperclone.exe
