+++ paperclone/lastnames.txt
@@ -1,0 +1,10 @@
+Smith
+Jones
+Brown
+Johnson
+Williams
+Miller
+Taylor
+Davis
+White
+Clark
