+++ paperclone/missions.txt
@@ -1,0 +1,6 @@
+Reconnaisance
+Movement
+Training
+Special Operations
+Peacekeeping
+Counter-Insurgency
