+++ paperclone/x64/Debug/paperclone.tlog/paperclone.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v140:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.16299.0
+Debug|x64|C:\Users\COOL USER\Source\Repos\PA9-Repository\|
